requirements.txt
crispy-bootstrap5==0.6
cryptography==38.0.1
Django==3.2
django-crispy-bootstrap==0.1.1.1
django-crispy-forms==1.14.0
django-faker==0.2
factory-boy==3.2.1
fake-factory==9999.9.9
Faker==15.0.0
fonttools==4.37.3
iniconfig==1.1.1
isort==5.10.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
nodejs==0.1.1
oauthlib==3.2.1
optional-django==0.1.0
psycopg2-binary==2.9.3
pycparser==2.21
pytest==7.1.3
pytest-django==4.5.2
pytest-factoryboy==2.5.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
widgetsnbextension==4.0.3

